requirements.txt
aniso8601==9.0.1
attrs==23.2.0
blinker==1.8.2
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.1
Flask==3.0.3
Flask-JWT-Extended==4.6.0
flask-marshmallow==1.2.1
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
importlib_resources==6.4.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
marshmallow==3.21.3
marshmallow-sqlalchemy==1.0.0
#mysqlclient==2.2.4
packaging==24.1
pluggy==1.5.0
Pygments==2.18.0
PyJWT==2.8.0
pytest==8.2.2
python-dotenv==1.0.1
pytz==2024.1
referencing==0.35.1
rpds-py==0.18.1
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.31
tomli==2.0.1
typing_extensions==4.12.2
Werkzeug==3.0.3
